+++ src/test/java/com/fakedevelopers/bidderbidder/service/UserServiceTest.java
@@ -1,0 +1,5 @@
+package com.fakedevelopers.bidderbidder.service;
+
+public class UserServiceTest {
+
+}
